+++ app/models/concerns/__init__.py
@@ -1,0 +1,5 @@
+from .events.attendable import Attendable
+from .events.collaborative import Collaborative
+
+
+__all__ = ["Attendable", "Collaborative"]
